fix: address MR review feedback
- Preserve null difficulty_value for bodyweight exercises instead of
  coercing to 0 (WorkoutEdit.vue, SetRow.vue)
- Change exercises.equipment_id FK from CASCADE to RESTRICT to prevent
  accidental data loss when deleting equipment
- Change difficulty_value column from integer to decimal(6,1) and align
  model cast (float), matching the frontend step=0.5 and numeric
  validation already in place
- Fix DifficultyUnit translation keys to use lowercase 'plates'
  matching lang/*.json keys so RU/other locales render correctly
- Add unit tests for DifficultyUnit and EffortType enums (100% patch
  coverage)

+++ app/Models/Set.php
@@ -24,7 +24,7 @@ class Set extends Model
             'activity_id' => 'integer',
             'order' => 'integer',
             'effort_value' => 'integer',
-            'difficulty_value' => 'integer',
+            'difficulty_value' => 'float',
             'is_completed' => 'boolean',
         ];
     }
